Move header text and trace up a bit
