+++ static/app/views/issueDetails/groupEventHeader.tsx
@@ -5,12 +5,9 @@ import GlobalAppStoreConnectUpdateAlert from 'sentry/components/globalAppStoreCo
 import {space} from 'sentry/styles/space';
 import {Group, Project} from 'sentry/types';
 import {Event} from 'sentry/types/event';
-import {useLocation} from 'sentry/utils/useLocation';
 import useOrganization from 'sentry/utils/useOrganization';
 import {GroupEventCarousel} from 'sentry/views/issueDetails/groupEventCarousel';
 import {OpenAIFixSuggestionPanel} from 'sentry/views/issueDetails/openAIFixSuggestion/openAIFixSuggestionPanel';
-
-import QuickTrace from './quickTrace';
 
 type GroupEventHeaderProps = {
   event: Event;
@@ -20,19 +17,12 @@ type GroupEventHeaderProps = {
 };
 
 const GroupEventHeader = ({event, group, project}: GroupEventHeaderProps) => {
-  const location = useLocation();
   const organization = useOrganization();
 
   return (
     <DataSection>
       <GroupEventCarousel group={group} event={event} projectSlug={project.slug} />
       <OpenAIFixSuggestionPanel projectSlug={project.slug} eventID={event.eventID} />
-      <QuickTrace
-        event={event}
-        group={group}
-        organization={organization}
-        location={location}
-      />
       <StyledGlobalAppStoreConnectUpdateAlert
         project={project}
         organization={organization}
