+++ static/app/views/issueDetails/quickTrace/index.tsx
@@ -1,11 +1,14 @@
 import {useContext} from 'react';
+import {useTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 import {Location} from 'history';
 
 import Placeholder from 'sentry/components/placeholder';
+import {space} from 'sentry/styles/space';
 import {Group, Organization} from 'sentry/types';
 import {Event} from 'sentry/types/event';
 import {QuickTraceContext} from 'sentry/utils/performance/quickTrace/quickTraceContext';
+import useMedia from 'sentry/utils/useMedia';
 
 import IssueQuickTrace from './issueQuickTrace';
 
@@ -17,11 +20,14 @@ type Props = {
 };
 
 function QuickTrace({group, event, organization, location}: Props) {
+  const theme = useTheme();
   const hasPerformanceView = organization.features.includes('performance-view');
   const hasTraceContext = Boolean(event.contexts?.trace?.trace_id);
   const quickTrace = useContext(QuickTraceContext);
 
-  if (!hasPerformanceView || !hasTraceContext) {
+  const isSmallViewport = useMedia(`(max-width: ${theme.breakpoints.small})`);
+
+  if (isSmallViewport || !hasPerformanceView || !hasTraceContext) {
     return null;
   }
 
@@ -43,6 +49,7 @@ function QuickTrace({group, event, organization, location}: Props) {
 const TracePlaceholder = styled(Placeholder)`
   width: auto;
   max-width: 300px;
+  margin-top: ${space(0.75)};
 `;
 
 export default QuickTrace;
